requirements.txt
asttokens==3.0.1
comm==0.2.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.0
fsspec==2026.2.0
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.8.0
jupyter-core==5.9.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.4
prompt-toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
setuptools==82.0.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
traitlets==5.14.3
triton==3.6.0
typing-extensions==4.15.0
wcwidth==0.6.0